Added JSON functionality to the PIN number - the user can now store their changed PIN number, and the program will prompt for a new PIN number upon opening if none exists in the JSON file.
Shifted the functions to the top of the code as there was an indentation issue meaning that if the incorrect PIN was entered in the beginning the program would still run. Fixed now.

I'm not convinced everything works perfectly just yet but will continue to test for issues - next up, making sure the PIN number is 4-digits (no more, no less)